requirements.txt
streamlit
pandas
pdfplumber
gspread
google-auth
plotly